requirements.txt
Flask==1.1.1
aspectlib==1.5.0
PyYAML==5.1.2